requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.28.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
